Refactor - Create a file for each class in utils
